Fix PHP 8.5 deprecation notices in the PayPal module (#3270)
* Replace non-canonical (double) casts in the PayPal module

PHP 8.5 deprecates the (double) cast alias and logs
"Non-canonical cast (double) is deprecated, use the (float) cast instead"
on every request that loads these files. Switch to the canonical (float)
cast, which is behaviourally identical.


* Mark IOException $previous parameter explicitly nullable

PHP 8.4 deprecates implicitly nullable parameters, so every request that
loads PayPalHttp\IOException logs a deprecation notice. Add the explicit
"?" to the type, which does not change the accepted values.


* update paypal - raise module version from 1.110 to 1.111 due to #3270


---------

+++ includes/external/paypal/classes/PayPalPayment.php
@@ -201,8 +201,8 @@ class PayPalPayment extends PayPalPaymentBase {
         $i = count($item);
         $item[$i] = $shipping_cost;
 
-        $this->amount->setTotal($this->amount->getTotal() + (double)$shipping_data['total'] + (double)$shipping_data['tax']);
-        $this->details->setTax($this->details->getTax() + (double)$shipping_data['tax']);
+        $this->amount->setTotal($this->amount->getTotal() + (float)$shipping_data['total'] + (float)$shipping_data['tax']);
+        $this->details->setTax($this->details->getTax() + (float)$shipping_data['tax']);
         $this->details->setSubtotal($this->amount->getTotal() - $this->details->getTax() + $this->details->getDiscount());
       }
       
@@ -294,7 +294,7 @@ class PayPalPayment extends PayPalPaymentBase {
     
       if (isset($shipping_cost) && is_object($shipping_cost)) {
         $item[1] = $shipping_cost;
-        $item[0]->setPrice($this->details->getSubtotal() - (double)$shipping_cost->getPrice());
+        $item[0]->setPrice($this->details->getSubtotal() - (float)$shipping_cost->getPrice());
       }    
     }
     $itemList->setItems($item);
